debug

+++ CycleGAN/setup.py
@@ -88,22 +88,22 @@ sample_zebra = next(iter(train_zebras))
 
 plt.subplot(121)
 plt.title('Horse')
-img_horse = float(sample_horse[0] * 0.5 + 0.5)
+img_horse = (sample_horse[0] * 0.5 + 0.5).astype(float)
 plt.imsave(opt.sample_data_path + 'sample_horse.jpg', img_horse, cmap=cmap)
 
 plt.subplot(122)
 plt.title('Horse with random jitter')
-img_horse_jitter = float(random_jitter(sample_horse[0]) * 0.5 + 0.5)
+img_horse_jitter = (random_jitter(sample_horse[0]) * 0.5 + 0.5).astype(float)
 plt.imsave(opt.sample_data_path + 'sample_horse_jitter.jpg', img_horse_jitter, cmap=cmap)
 
 plt.subplot(121)
 plt.title('Zebra')
-img_zebra = float(sample_zebra[0] * 0.5 + 0.5)
+img_zebra = (sample_zebra[0] * 0.5 + 0.5).astype(float)
 plt.imsave(opt.sample_data_path + 'sample_zebra.jpg', img_zebra, cmap=cmap)
 
 plt.subplot(122)
 plt.title('Zebra with random jitter')
-img_zebra_jitter = float(random_jitter(sample_zebra[0]) * 0.5 + 0.5)
+img_zebra_jitter = (random_jitter(sample_zebra[0]) * 0.5 + 0.5).astype(float)
 plt.imsave(opt.sample_data_path + 'sample_zebra.jpg', img_zebra_jitter, cmap=cmap)
 plt.show()
